refactor(runtime): extract history-compact engine from context-budget
Move the history-compact block engine out of context-budget.ts into
history-compact.ts, keeping behavior neutral:

- extract ~967 lines (31 symbols: applyRuntimeEventHistoryCompact, the
  build/render/validate/select history-compact block helpers, coverage,
  deterministic summary, checkpoint replay gate, isHistoryCompactContentEvent,
  replaceHistoryCompactReplayBlocks) to history-compact.ts
- break the ContextBudgetPolicy reverse import by narrowing
  applyRuntimeEventHistoryCompact and evaluateHistoryCompactCheckpointReplay
  to accept the compact sub-policy + charsPerToken + maxHistoryEstimatedTokens
  instead of the whole ContextBudgetPolicy
- relocate the cross-block pure helpers finiteRatio/normalizeWhitespace/
  boundText to context-budget-helpers.ts (now shared by history-compact and
  the budget surface)
- re-export all public symbols from context-budget.ts for import
  compatibility and adapt the 5 call-site tests to the narrowed signatures

No reverse imports: history-compact only imports the leaves
context-budget-helpers, synthesis-cache, compaction-boundary,
history-compact-checkpoint, and @maka/core; context-budget and ai-sdk-backend
import forward.

Part of #1084.
