+++ apps/desktop/src/main/__tests__/history-compact-artifacts.test.ts
@@ -232,15 +232,12 @@ describe('desktop history compact artifact lifecycle', () => {
         now: () => 1_800_000_000_100,
         summarize: () => 'host summary alpha beta',
       });
-      const replay = applyRuntimeEventHistoryCompact([
-        ...foldedEvents,
-        textEvent('recent', 'turn-3', 'recent retained fact'),
-      ], {
-        name: 'archive-required-replay',
-        maxHistoryEstimatedTokens: 2_048,
-        minRecentTurns: 1,
-        charsPerToken: 4,
-        historyCompact: {
+      const replay = applyRuntimeEventHistoryCompact(
+        [
+          ...foldedEvents,
+          textEvent('recent', 'turn-3', 'recent retained fact'),
+        ],
+        {
           enabled: true,
           mode: 'lookup',
           highWaterRatio: 0.000001,
@@ -249,7 +246,9 @@ describe('desktop history compact artifact lifecycle', () => {
           archiveRequired: true,
           blocks: write.blocks,
         },
-      });
+        4,
+        2_048,
+      );
 
       assert.equal(replay.blocks.length, 1);
       assert.equal(replay.blocks[0]?.sourceArchiveRefs?.length, 2);
